+++ src/main/lifeCycle/index.ts
@@ -41,18 +41,20 @@ const isDevelopment = process.env.NODE_ENV !== 'production'
 
 const handleStartUpFiles = (argv: string[], cwd: string) => {
   const files = getUploadFiles(argv, cwd, logger)
-  if (files === null || files.length > 0) {
-    // 如果有文件列表作为参数，说明是命令行启动
-    if (files === null) {
-      logger.info('cli -> uploading file from clipboard')
-      uploadClipboardFiles()
-    } else {
-      logger.info('cli -> uploading files from cli', ...files.map(item => item.path))
-      const win = windowManager.getAvailableWindow()
-      uploadChoosedFiles(win.webContents, files)
-    }
+
+  if (files === null) {
+    logger.info('cli -> uploading file from clipboard')
+    uploadClipboardFiles()
     return true
   }
+
+  if (files.length > 0) {
+    logger.info('cli -> uploading files from cli', ...files.map(file => file.path))
+    const win = windowManager.getAvailableWindow()
+    uploadChoosedFiles(win.webContents, files)
+    return true
+  }
+
   return false
 }