+++ src/renderer/manage/components/DynamicSwitch.vue
@@ -5,7 +5,7 @@
         <span v-for="(segment, index) in segments" :key="index" :style="segment.style">
           {{ segment.text }}
         </span>
-        <el-tooltip :content="tooltip" effect="dark" placement="right" :persistent="false" teleported>
+        <el-tooltip v-if="tooltip" :content="tooltip" effect="dark" placement="right" :persistent="false" teleported>
           <el-icon>
             <InfoFilled />
           </el-icon>